steering wheel modes added
Two additional models were added to the noise simulation stage, one of which restricted the angle change to +- a user-defined value. The model that uses a first-order transfer function with a P controller was added to limit the angle change of the steering wheel. The model that uses a T1FLS with a Proportional controller was added. Both models work at 20ms and the main loop at 100ms, which is why they are executed inside a for loop 5 times before the main loop.